feat: keep the half hour the four practices print (#28)
Weekly hours were a whole number, so the four items the published plans print with a
half were stored rounded up: Psicología's two professional practices at 4,5 hours, and
the practices of Marketing and Negocios Internacionales at 1,5.

That rounding is what made Psicología add up to 175 weekly hours against the 174 its own
document prints - a plan disagreeing with its own printed total for a reason that was
ours, not the document's. Kept as printed, its items add up to exactly 174. Its printed
per-semester totals still say 175 because the document rounds each half up for display
(14,5 printed as 15 and 10,5 as 11), so both printed figures turn out to be right, each
in its own way. PREGUNTAS-PENDIENTES.md question 6 is answered; question 4 keeps the one
thing only the university can confirm, that the halves are real.

weeklyHours, a pensum's totalHours and an area's hours are decimals now, and halves are
the only fraction allowed: 4.3 is refused as the transcription slip it would be. A
course's totalHours stays a whole number, because half an hour a week is eight whole
hours a semester.

Whole hours are still written whole. A serialiser emits 4 rather than 4.0, so those four
items are the only place in the API where hours carry a decimal point and the payloads
for the other 646 are byte-for-byte what they were - which matters while the Kotlin and
Swift clients are being written against them. Those clients do need a decimal type for
the four.

The CSV import takes 4.5 or the Spanish 4,5, quoted - unquoted it is not a decimal mark
but the next column, which Commons CSV catches first. The paste importer says the same
before the file is built, rather than at the end of the import.

V007 moves the four in databases that already ran V006, the shared cluster among them,
and only while each still holds exactly the rounded figure, so an item somebody has
corrected by hand is left alone. Which plans declare their own total is written down
rather than inferred: after the fix the figures agree, and a rule that read "the total
equals its items, so it was derived" would push Psicología's printed 174 to 175 on the
way back.

+++ app/backend/microservices/semaphore-service/src/main/java/co/edu/konradlorenz/kapp/semaphore/web/dto/PensumImportReport.java
@@ -30,8 +30,8 @@ public record PensumImportReport(
             int courses,
             int declaredCredits,
             int computedCredits,
-            int declaredHours,
-            int computedHours,
+            double declaredHours,
+            double computedHours,
             boolean programCreated,
             boolean pensumCreated
     ) {
